requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
cachetools==3.0.0
cffi==1.11.5
Click==7.0
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
Flask==1.0.2
Flask-Cors==3.0.7
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10
joblib==0.13.0
lxml==4.2.5
MarkupSafe==1.1.0
numpy==1.15.4
pandas==0.23.4
parsel==1.5.1
psycopg2==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
queuelib==1.5.0
scikit-surprise==1.0.6
scipy==1.1.0
Scrapy==1.5.1
selenium==3.141.0
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.14
Twisted==18.9.0
urllib3==1.24.1
w3lib==1.19.0
Werkzeug==0.14.1
zope.interface==4.6.0
